refactor(markdown): extract components and make them display
